feat(providers): stream gemini generateContent responses
