+++ docs/examples/provider.env.example
@@ -4,3 +4,4 @@ DEEPSEEK_API_KEY=
 MIMO_API_KEY=
 LOCAL_OPENAI_COMPAT_API_KEY=
 ANTHROPIC_API_KEY=
+GEMINI_API_KEY=
