+++ src/providers/gemini-generate-content/types.ts
@@ -1,0 +1,38 @@
+export type GeminiPart = {
+  text?: string;
+  thought?: boolean;
+  thoughtSignature?: string;
+  functionCall?: {
+    id?: string;
+    name?: string;
+    args?: unknown;
+  };
+  functionResponse?: {
+    id?: string;
+    name?: string;
+    response?: unknown;
+  };
+  [key: string]: unknown;
+};
+
+export type GeminiContent = {
+  role?: "user" | "model";
+  parts?: GeminiPart[];
+};
+
+export type GeminiCandidate = {
+  content?: GeminiContent;
+  finishReason?: string;
+};
+
+export type GeminiResponse = {
+  candidates?: GeminiCandidate[];
+  usageMetadata?: {
+    promptTokenCount?: number;
+    candidatesTokenCount?: number;
+    totalTokenCount?: number;
+  };
+  error?: {
+    message?: string;
+  };
+};
